Split diagnostics.rs into multiple files
