+++ compiler/plc_diagnostics/src/diagnostics.rs
@@ -5,15 +5,15 @@ use std::{
     ops::Range,
 };
 
-use codespan_reporting::{
-    diagnostic::Label,
-    files::{Files, Location, SimpleFile, SimpleFiles},
-    term::termcolor::{ColorChoice, StandardStream},
-};
-// use inkwell::support::LLVMString;
 use plc_ast::ast::{AstStatement, DataTypeDeclaration, DiagnosticInfo, PouType, SourceRange};
 
-// use crate::index::VariableType;
+use crate::{
+    errno::ErrNo,
+    reporter::{
+        clang::ClangFormatDiagnosticReporter, codespan::CodeSpanDiagnosticReporter,
+        null::NullDiagnosticReporter,
+    },
+};
 
 pub const INTERNAL_LLVM_ERROR: &str = "internal llvm codegen error";
 
@@ -24,94 +24,6 @@ pub enum Diagnostic {
     GeneralError { message: String, err_no: ErrNo },
     ImprovementSuggestion { message: String, range: Vec<SourceRange> },
     CombinedDiagnostic { message: String, inner_diagnostics: Vec<Diagnostic>, err_no: ErrNo },
-}
-
-#[allow(non_camel_case_types)]
-#[derive(PartialEq, Eq, Debug, Clone, Copy)]
-pub enum ErrNo {
-    undefined,
-
-    //general
-    general__io_err,
-    general__param_err,
-    duplicate_symbol,
-
-    //syntax
-    syntax__generic_error,
-    syntax__missing_token,
-    syntax__unexpected_token,
-
-    //semantic
-    // pou related
-    pou__missing_return_type,
-    pou__unexpected_return_type,
-    pou__unsupported_return_type,
-    pou__empty_variable_block,
-    pou__missing_action_container,
-    pou__recursive_data_structure,
-
-    // call
-    call__invalid_parameter_type,
-    call__invalid_parameter_count,
-
-    //variable related
-    var__unresolved_constant,
-    var__invalid_constant_block,
-    var__invalid_constant,
-    var__cannot_assign_to_const,
-    var__invalid_assignment,
-    var__missing_type,
-    var__assigning_to_var_input_ref,
-    var__overflow,
-    var__invalid_enum_variant,
-
-    //array related
-    arr__invalid_array_assignment,
-
-    // VLA related
-    vla__invalid_container,
-    vla__invalid_array_access,
-    vla__dimension_idx_out_of_bounds,
-
-    //reference related
-    reference__unresolved,
-    reference__illegal_access,
-    reference__expected,
-
-    //type related
-    type__cast_error,
-    type__unknown_type,
-    type__invalid_type,
-    type__literal_out_of_range,
-    type__incompatible_literal_cast,
-    type__incompatible_directaccess,
-    type__incompatible_directaccess_variable,
-    type__incompatible_directaccess_range,
-    type__incompatible_arrayaccess_range,
-    type__incompatible_arrayaccess_variable,
-    type__incompatible_arrayaccess_type,
-    type__expected_literal,
-    type__invalid_nature,
-    type__unknown_nature,
-    type__unresolved_generic,
-    type__incompatible_size,
-    type__invalid_operation,
-    type__invalid_name,
-
-    //codegen related
-    codegen__general,
-    codegen__missing_function,
-    codegen__missing_compare_function,
-
-    //Debug code
-    debug_general,
-    //linker
-    linker__generic_error,
-
-    //switch case
-    case__duplicate_condition,
-    case__case_condition_outside_case_statement,
-    case__invalid_case_condition,
 }
 
 impl<T: Error> From<T> for Diagnostic {
@@ -888,183 +800,6 @@ pub trait DiagnosticReporter {
     /// a relationship the given src (diagnostics for this src need
     /// to use this id)
     fn register(&mut self, path: String, src: String) -> usize;
-}
-
-/// a DiagnosticReporter that reports diagnostics using codespan_reporting
-pub struct CodeSpanDiagnosticReporter {
-    files: SimpleFiles<String, String>,
-    config: codespan_reporting::term::Config,
-    writer: StandardStream,
-}
-
-impl CodeSpanDiagnosticReporter {
-    /// creates a new reporter with the given codespan_reporting configuration
-    fn new(config: codespan_reporting::term::Config, writer: StandardStream) -> Self {
-        CodeSpanDiagnosticReporter { files: SimpleFiles::new(), config, writer }
-    }
-}
-
-impl Default for CodeSpanDiagnosticReporter {
-    /// creates the default CodeSpanDiagnosticReporter reporting to StdErr, with colors
-    fn default() -> Self {
-        Self::new(
-            codespan_reporting::term::Config {
-                display_style: codespan_reporting::term::DisplayStyle::Rich,
-                tab_width: 2,
-                styles: codespan_reporting::term::Styles::default(),
-                chars: codespan_reporting::term::Chars::default(),
-                start_context_lines: 5,
-                end_context_lines: 3,
-            },
-            StandardStream::stderr(ColorChoice::Always),
-        )
-    }
-}
-
-impl DiagnosticReporter for CodeSpanDiagnosticReporter {
-    fn report(&self, diagnostics: &[ResolvedDiagnostics]) {
-        for d in diagnostics {
-            let diagnostic_factory = match d.severity {
-                Severity::Error => codespan_reporting::diagnostic::Diagnostic::error(),
-                Severity::Warning => codespan_reporting::diagnostic::Diagnostic::warning(),
-                Severity::_Info => codespan_reporting::diagnostic::Diagnostic::note(),
-            };
-
-            let mut labels = vec![Label::primary(d.main_location.file_handle, d.main_location.range.clone())
-                .with_message(d.message.as_str())];
-
-            if let Some(additional_locations) = &d.additional_locations {
-                labels.extend(
-                    additional_locations.iter().map(|it| {
-                        Label::secondary(it.file_handle, it.range.clone()).with_message("see also")
-                    }),
-                );
-            }
-
-            let diag = diagnostic_factory.with_labels(labels).with_message(d.message.as_str());
-
-            let result =
-                codespan_reporting::term::emit(&mut self.writer.lock(), &self.config, &self.files, &diag);
-            if result.is_err() && d.main_location.is_internal() {
-                eprintln!("<internal>: {}", d.message);
-            }
-        }
-    }
-
-    fn register(&mut self, path: String, src: String) -> usize {
-        self.files.add(path, src)
-    }
-}
-
-/// a DiagnosticReporter that reports diagnostics using clang-format
-pub struct ClangFormatDiagnosticReporter {
-    files: SimpleFiles<String, String>,
-}
-
-impl ClangFormatDiagnosticReporter {
-    fn new() -> Self {
-        ClangFormatDiagnosticReporter { files: SimpleFiles::new() }
-    }
-}
-
-impl Default for ClangFormatDiagnosticReporter {
-    fn default() -> Self {
-        ClangFormatDiagnosticReporter::new()
-    }
-}
-
-impl DiagnosticReporter for ClangFormatDiagnosticReporter {
-    fn report(&self, diagnostics: &[ResolvedDiagnostics]) {
-        for diagnostic in diagnostics {
-            let file_id = diagnostic.main_location.file_handle;
-            let location = &diagnostic.main_location;
-
-            let file = self.files.get(file_id).ok();
-            let start = self.files.location(file_id, location.range.start).ok();
-            let end = self.files.location(file_id, location.range.end).ok();
-
-            let res = self.build_diagnostic_msg(
-                file,
-                start.as_ref(),
-                end.as_ref(),
-                &diagnostic.severity,
-                &diagnostic.message,
-            );
-
-            eprintln!("{res}");
-        }
-    }
-    fn register(&mut self, path: String, src: String) -> usize {
-        self.files.add(path, src)
-    }
-}
-
-impl ClangFormatDiagnosticReporter {
-    /// returns diagnostic message in clang format
-    /// file-name:{range}: severity: message
-    /// optional parameters that are none will not be included
-    fn build_diagnostic_msg(
-        &self,
-        file: Option<&SimpleFile<String, String>>,
-        start: Option<&Location>,
-        end: Option<&Location>,
-        severity: &Severity,
-        msg: &str,
-    ) -> String {
-        let mut str = String::new();
-        // file name
-        if let Some(f) = file {
-            str.push_str(format!("{}:", f.name().as_str()).as_str());
-            // range
-            if let Some(s) = start {
-                if let Some(e) = end {
-                    // if start and end are equal there is no need to show the range
-                    if s.eq(e) {
-                        str.push_str(format!("{}:{}: ", s.line_number, s.column_number).as_str());
-                    } else {
-                        str.push_str(
-                            format!(
-                                "{}:{}:{{{}:{}-{}:{}}}: ",
-                                s.line_number,
-                                s.column_number,
-                                s.line_number,
-                                s.column_number,
-                                e.line_number,
-                                e.column_number
-                            )
-                            .as_str(),
-                        );
-                    }
-                }
-            } else {
-                str.push(' ');
-            }
-        }
-        // severity
-        str.push_str(format!("{severity}: ").as_str());
-        // msg
-        str.push_str(msg);
-
-        str
-    }
-}
-
-/// a DiagnosticReporter that swallows all diagnostics
-#[derive(Default)]
-pub struct NullDiagnosticReporter {
-    last_id: usize,
-}
-
-impl DiagnosticReporter for NullDiagnosticReporter {
-    fn report(&self, _diagnostics: &[ResolvedDiagnostics]) {
-        //ignore
-    }
-
-    fn register(&mut self, _path: String, _src: String) -> usize {
-        // at least provide some unique ids
-        self.last_id += 1;
-        self.last_id
-    }
 }
 
 /// the Diagnostician handle's Diangostics with the help of a
