+++ compiler/plc_diagnostics/src/lib.rs
@@ -1,6 +1,9 @@
-//! This crates handles diagnostics, from creation
-//! to reporting.
-//! It is currently only a re-export of the diagnostics functions from the root, but these should
-//! eventually move here
+//! This crates handles diagnostics, from creation to reporting.
 
 pub mod diagnostics;
+pub mod reporter {
+    pub mod clang;
+    pub mod codespan;
+    pub mod null;
+}
+pub mod errno; // TODO: Make this crate-private?
